workers/test_client/requirements.txt
paho-mqtt
redis
influxdb-client